refactor, report: migrating filter formatting functions to utils, implement filtering into html report.
