+++ lib/report/html-template.js
@@ -7,7 +7,7 @@ const {
 
 module.exports = renderHTML
 
-function renderHTML (title, summary, report, whitelist) {
+function renderHTML (title, summary, report, reportLength, whitelist, whitelistLength, formattedFilterOptions) {
   const { riskCount, securityCount, insecureModules, complianceCount } = summary
 
   let alternate = false
@@ -162,7 +162,7 @@ function renderHTML (title, summary, report, whitelist) {
 
       <div id="summary" class="bottom-20 left-20">
         <span class="subtitle">Summary</span>
-        <p><b class="white">${report.length}</b> packages checked</p>
+        <p><b class="white">${reportLength}</b> packages checked</p>
         <br>
         <p><span class="red"><b>${riskCount[4]}</b> Critical Risk</span></p>
         <p><span class="orange"><b>${riskCount[3]}</b> High Risk</span></p>
@@ -181,15 +181,27 @@ function renderHTML (title, summary, report, whitelist) {
     ? `<b class="white">${complianceCount}</b> noncompliant modules found`
     : '<span class="green">✓</span> All modules compliant')}
         </p>
-    ${(whitelist.length > 0
-    ? `<p><b class="white">${whitelist.length}</b> used modules whitelisted</p>`
+    ${(whitelistLength > 0
+    ? `<p><b class="white">${whitelistLength}</b> used modules whitelisted</p>`
     : '')}
       </div>
 
-      ${(whitelist.length > 0
+      ${(whitelistLength > 0
     ? `
           <div id="whitelist-list">
-            <span class="subtitle">Whitelisted Modules</span>
+            <p>
+              <span class="subtitle">
+                Whitelisted
+    ${(formattedFilterOptions.length > 9
+      ? 'Filtered'
+      : '')}
+                Modules
+              </span>
+    ${(formattedFilterOptions.length > 9
+      ? `(${formattedFilterOptions})`
+      : '')}
+
+            </p>
             <hr>
             <table id="whitelist-table" class="bottom-20">
               <tr style="margin-bottom:5px;">
@@ -205,7 +217,20 @@ function renderHTML (title, summary, report, whitelist) {
     : '')}
       
       <div id="module-list">
-        <p class="subtitle">${(whitelist.length > 0 ? 'Non-Whitelisted ' : '')}Modules</p>
+        <p>
+          <span class="subtitle">
+    ${(whitelistLength > 0
+    ? 'Non-Whitelisted'
+    : '')}
+    ${(formattedFilterOptions.length > 9
+    ? 'Filtered'
+    : '')}
+            Modules
+          </span>
+    ${(formattedFilterOptions.length > 9
+    ? `( ${formattedFilterOptions} )`
+    : '')}
+        </p>
         <hr>
         <table id="module-table" class="bottom-20">
           <tr style="margin-bottom:5px;">
