+++ lib/report/html.js
@@ -8,14 +8,18 @@ const {
   formatError
 } = require('../ncm-style')
 const {
-  summaryInfo
+  summaryInfo,
+  moduleSort,
+  filterReport,
+  parseFilterOptions,
+  formatFilterOptions
 } = require('./util')
 const renderTemplate = require('./html-template')
 const L = console.log
 
 module.exports = htmlReport
 
-async function htmlReport (report, whitelist, dir, output) {
+async function htmlReport (report, whitelist, dir, output, argv) {
   /* Output may only use the `.html` file format */
   if (output !== true && !(/^.*\.html$/.test(path.basename(output)))) {
     L()
@@ -27,12 +31,25 @@ async function htmlReport (report, whitelist, dir, output) {
 
   const title = `${path.basename(dir) || 'NCM'}`
   const summary = summaryInfo(report) // { riskCount, insecureModules, complianceCount, securityCount }
+  const reportLength = report.length
+  const whitelistLength = whitelist.length
+
+  const filterOptions = parseFilterOptions(argv)
+  const formattedfilterOptions = formatFilterOptions(filterOptions)
+  report = filterReport(report, filterOptions)
+  report = moduleSort(report)
+
+  whitelist = filterReport(whitelist, filterOptions)
+  whitelist = moduleSort(whitelist)
 
   const htmlData = renderTemplate(
     title,
     summary,
     report,
-    whitelist
+    reportLength,
+    whitelist,
+    whitelistLength,
+    formattedfilterOptions
   )
 
   /*
