+++ lib/report/util.js
@@ -42,7 +42,10 @@ module.exports = {
   moduleList,
   moduleSort,
   summaryInfo,
-  filterVulns
+  filterVulns,
+  filterReport,
+  parseFilterOptions,
+  formatFilterOptions
 }
 
 function filterReport (report, options) {
@@ -287,4 +290,59 @@ function filterVulns (failures) {
   }
 
   return vulns
+}
+
+function parseFilterOptions (argv) {
+  let filterSecurity = argv ? !!argv.security : false
+  let filterCompliance = argv ? !!argv.compliance : false
+  let filterLevel = SEVERITY_RMAP.indexOf('NONE')
+
+  if (argv.filter) {
+    const segments = argv.filter.split(',')
+      .map(s => s.trim().toLowerCase())
+
+    if (segments.includes('compliance')) {
+      filterCompliance = true
+    }
+    if (segments.includes('security')) {
+      filterSecurity = true
+    }
+    if (segments.includes('c') || segments.includes('critical')) {
+      filterLevel = SEVERITY_RMAP.indexOf('CRITICAL')
+    }
+    if (segments.includes('h') || segments.includes('high')) {
+      filterLevel = SEVERITY_RMAP.indexOf('HIGH')
+    }
+    if (segments.includes('m') || segments.includes('medium')) {
+      filterLevel = SEVERITY_RMAP.indexOf('MEDIUM')
+    }
+    if (segments.includes('l') || segments.includes('low')) {
+      filterLevel = SEVERITY_RMAP.indexOf('LOW')
+    }
+  }
+
+  const filterOptions = {
+    filterCompliance: filterCompliance,
+    filterSecurity: filterSecurity,
+    filterLevel: filterLevel
+  }
+
+  return filterOptions
+}
+
+function formatFilterOptions (filterOptions) {
+  let str = '--filter='
+  if (filterOptions.filterCompliance) {
+    str += 'compliance,'
+  }
+  if (filterOptions.filterSecurity) {
+    str += 'security,'
+  }
+  if (filterOptions.filterLevel) {
+    str += SEVERITY_RMAP[filterOptions.filterLevel] + ','
+  }
+  if (str[str.length - 1] === ',') {
+    str = str.slice(0, str.length - 1)
+  }
+  return str
 }
